feat: implement preliminary joystick support + make WheelMovementPacket use bytes instead of floats
- basic joystick support has been added for both Desktop and Android platforms
- WheelMovementPacket uses bytes instead of floats for efficiency. This will require a code change on the receiving side of the serial connection as well.

+++ shared/src/commonMain/kotlin/me/arianb/storm_robot/SerializableDataClasses.kt
@@ -4,9 +4,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class WheelMovementPacket(
-    val x: Float,
-    val y: Float,
-    val rotation: Float
+    val x: Byte,
+    val y: Byte,
+    val rotation: Byte
 )
 
 @Serializable
